pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "nomad_ppms_plugin"
description = "Can read and analyze PPMS files"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Jonathan Noky", email = "jonathan.noky@cpfs.mpg.de" },
]
maintainers = [
    { name = "Jonathan Noky", email = "jonathan.noky@cpfs.mpg.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.0",
    "python-magic-bin; sys_platform == 'win32'",
    "nomad-measurements",
]

[project.urls]
Repository = "https://github.com/MPI-CPfS-Dresden/nomad_ppms_plugin"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
eto_parser_default = "nomad_ppms_plugin:eto_parser_default"
eto_parser_labview = "nomad_ppms_plugin:eto_parser_labview"
act_parser_default = "nomad_ppms_plugin:act_parser_default"
mpms_parser_default = "nomad_ppms_plugin:mpms_parser_default"
acms_parser_default = "nomad_ppms_plugin:acms_parser_default"
resistivity_parser_default = "nomad_ppms_plugin:resistivity_parser_default"
eto_schema_default = "nomad_ppms_plugin:eto_schema_default"
eto_schema_labview = "nomad_ppms_plugin:eto_schema_labview"
act_schema_default = "nomad_ppms_plugin:act_schema_default"
mpms_schema_default = "nomad_ppms_plugin:mpms_schema_default"
acms_schema_default = "nomad_ppms_plugin:acms_schema_default"
resistivity_schema_default = "nomad_ppms_plugin:resistivity_schema_default"



[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
